Rename library to module as Swift seems to prefer that term
