+++ subprojects/language-native/src/integTest/groovy/org/gradle/language/swift/plugins/SwiftModulePluginIntegrationTest.groovy
@@ -16,23 +16,5 @@
 
 package org.gradle.language.swift.plugins
 
-import org.gradle.language.swift.tasks.SwiftCompile
-import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider
-import org.gradle.util.TestUtil
-import org.junit.Rule
-import spock.lang.Specification
-
-class SwiftLibraryPluginTest extends Specification {
-    @Rule
-    TestNameTestDirectoryProvider tmpDir = new TestNameTestDirectoryProvider()
-    def project = TestUtil.createRootProject(tmpDir.createDir("project"))
-
-    def "adds compile and link tasks"() {
-        when:
-        project.pluginManager.apply(SwiftLibraryPlugin)
-
-        then:
-        def compileSwift = project.tasks.compileSwift
-        compileSwift instanceof SwiftCompile
-    }
+class SwiftModulePluginIntegrationTest extends AbstractSwiftPluginIntegrationTest {
 }
