+++ subprojects/language-native/src/main/java/org/gradle/language/swift/plugins/SwiftModulePlugin.java
@@ -44,7 +44,7 @@ import java.util.Collections;
  * @since 4.1
  */
 @Incubating
-public class SwiftLibraryPlugin implements Plugin<Project> {
+public class SwiftModulePlugin implements Plugin<Project> {
     @Override
     public void apply(Project project) {
         project.getPluginManager().apply(SwiftBasePlugin.class);
