+++ subprojects/language-native/src/main/resources/META-INF/gradle-plugins/org.gradle.swift-module.properties
@@ -14,4 +14,4 @@
 # limitations under the License.
 #
 
-implementation-class=org.gradle.language.swift.plugins.SwiftLibraryPlugin
+implementation-class=org.gradle.language.swift.plugins.SwiftModulePlugin
